+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - GCGI-186: Setup mode in main script
+- GCGI-187: Fail and target-coverage options for HTML generation
 
 ## v0.0.6: 2021-08-19